ref: Fix tracing bundle build without replay
